feat: update structure using the terminology suggest by Gavin
